+++ ChangeLog.md
@@ -2,7 +2,7 @@
 
 ## NEXT
 
-  *
+  * Make sure command line --config options get turned from strings into booleans if needed
 
 ## v2.0.2 (2012-09-19)